+++ runtime/src/lib.rs
@@ -7,6 +7,7 @@ pub mod apis;
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarks;
 pub mod configs;
+pub mod inherent_checks;
 
 extern crate alloc;
 use alloc::vec::Vec;
